+++ lib/src/interfaces/connection_listener.dart
@@ -1,0 +1,3 @@
+class ConnectionListener {
+  void onResponse(bool internetAvailable) {}
+}