+++ lib/src/widgets/custom_button.dart
@@ -4,22 +4,30 @@ import 'package:sizer/sizer.dart';
 
 import '../helpers/constants.dart';
 import '../helpers/styles.dart';
-
 class CustomButton extends StatefulWidget {
   Color? color;
   String text;
   double? width;
   double? height;
   EdgeInsetsGeometry? margin;
+  EdgeInsetsGeometry? padding;
   VoidCallback onPressed;
+  VoidCallback? onLongPressed;
+  TextStyle? textStyle;
+  TextAlign? textAlign;
+  OutlinedBorder? shape;
 
   CustomButton(
       {this.color,
-      required this.text,
-      required this.onPressed,
-      this.width,
-      this.margin,
-      this.height});
+        required this.text,
+        required this.onPressed,
+        this.width,
+        this.margin,
+        this.height,
+        this.textStyle,
+        this.shape,
+        this.padding,
+        this.textAlign, this.onLongPressed});
 
   @override
   _CustomButtonState createState() => _CustomButtonState();
@@ -31,22 +39,25 @@ class _CustomButtonState extends State<CustomButton> {
     return GestureDetector(
       onTap: widget.onPressed,
       child: Container(
-        margin:
-            widget.margin ?? EdgeInsets.symmetric(vertical: 20, horizontal: 10),
-        height: widget.height ?? 7.h,
-        width: widget.width ?? MediaQuery.of(context).size.width,
+        margin: widget.margin ??
+            const EdgeInsets.symmetric(vertical: 20, horizontal: 10),
+        height: widget.height,
+        width: widget.width,
         child: ElevatedButton(
           onPressed: widget.onPressed,
+          onLongPress: widget.onLongPressed,
           style: ElevatedButton.styleFrom(
-              padding: EdgeInsets.all(10),
-              primary: widget.color ?? Colors.red,
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.all(Radius.circular(10)),
-              )),
+              padding: widget.padding,
+              primary: widget.color ?? buttonColor,
+              shape: widget.shape ??
+                  const RoundedRectangleBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(10)),
+                  )),
           child: Text(
             widget.text,
-            textAlign: TextAlign.center,
-            style: normal_h3Style.copyWith(color: Colors.white),
+            textAlign: widget.textAlign ?? TextAlign.center,
+            style: widget.textStyle ??
+                normal_h3Style.copyWith(color: Colors.white),
           ),
         ),
       ),