+++ lib/src/widgets/custom_listview_builder.dart
@@ -10,34 +10,27 @@ class CustomListviewBuilder extends StatefulWidget {
   @override
   _CustomListviewBuilderState createState() => _CustomListviewBuilderState();
 
-  CustomListviewBuilder(
-      {required this.itemBuilder,
-      required this.itemCount,
-      required this.scrollDirection,
-      this.scrollable,
-      this.reverse});
+  CustomListviewBuilder({required this.itemBuilder, required this.itemCount, required this.scrollDirection, this.scrollable, this.reverse});
 }
 
 class _CustomListviewBuilderState extends State<CustomListviewBuilder> {
   @override
   Widget build(BuildContext context) {
-    return SingleChildScrollView(
-      reverse: widget.reverse ?? false,
-      physics: (widget.scrollable != null && widget.scrollable == false)
-          ? NeverScrollableScrollPhysics()
-          : BouncingScrollPhysics(),
-      scrollDirection: widget.scrollDirection == CustomDirection.vertical
-          ? Axis.vertical
-          : Axis.horizontal,
-      child: widget.scrollDirection == CustomDirection.vertical
-          ? Column(
-              children: List.generate(widget.itemCount,
-                  (index) => widget.itemBuilder(context, index)).toList(),
-            )
-          : Row(
-              children: List.generate(widget.itemCount,
-                  (index) => widget.itemBuilder(context, index)).toList(),
-            ),
+    return Container(
+      alignment: Alignment.topLeft,
+      child: SingleChildScrollView(
+        reverse: widget.reverse ?? false,
+        physics: (widget.scrollable != null && widget.scrollable == false) ? NeverScrollableScrollPhysics() : BouncingScrollPhysics(),
+        scrollDirection: widget.scrollDirection == CustomDirection.vertical ? Axis.vertical : Axis.horizontal,
+        child: widget.scrollDirection == CustomDirection.vertical
+            ? Column(
+          children: List.generate(widget.itemCount, (index) => widget.itemBuilder(context, index)).toList(),
+        )
+            : Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: List.generate(widget.itemCount, (index) => widget.itemBuilder(context, index)).toList(),
+        ),
+      ),
     );
   }
 }