+++ CHANGELOG.md
@@ -8,7 +8,9 @@
 * Fixed imports bugs
 ## 0.0.15
 * Fixed Async Validator in TextField (CustomInputField)
-## 0.0.16
+## 0.0.18
 * Added CustomInputController in CustomInputField
   - Now you can use custom mask/pattern for CustomInputField using "mask" in CustomInputController (e.g. 00000-0000000-0)
-  - Note: You can use a one controller at once (either TextEditingController or CustomInputController)
+  - Note: You can use only one controller at once (either TextEditingController or CustomInputController)
+## 0.0.21
+* Bugs Fixes